requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
pycryptodome==3.19.0
bcrypt==4.1.1
Werkzeug==3.0.1
